YD-143 Separating Spring-based test and unit test
